device: simplify changing device up/down state
The existing code had two different bools tracking up/down state:
device.isUp and device.state.current.

isUp is widely used; it indicated the current *or* aspirational state of the device.
Most uses of device.isUp assume that it represents the current state.
device.state.current was only used while the device's state was changing.

This commit aims to simplify the code.

It tracks the aspirational state of the device with a local variable.
To avoid bare "true" and "false", which are hard to follow at call sites,
it introduces a named type with mnemonic constants.
It makes isUp represent the actual state of the device.

This commit also aims to make state changes more atomic.
There was a short period when calling device.Up and device.Down
during which device.isUp did not match the device's state.
If device.BindUpdate failed, there was a short period before the
recursive call to deviceUpdateState during which device.state.mu
was unlocked and device.state.changing was false.
This commit fixes these (minuscule) logical races.

The reason I want to simplify and solidify this code is that on a weekend-long
stress test I saw three deadlocks in the device package tests
(out of 7.5 million runs). The deadlocks relate somehow to device.Up
and device.Down. The unordered, racy updates of device.isUp
and device.state.current made it hard to reason about the state machine,
and the stack traces didn't provide many hints.

I hope that this commit will either fix those deadlocks
or make them easier to diagnose.

Signed-off-by: Josh Bleecher Snyder <josh@tailscale.com>